+++ changelog.md
@@ -4,12 +4,11 @@
 *estimated May 2020*
 
 * rules: allow farmer placement on field with shepherd (#305)
+* rules: Shepherd is automatically returned when farm is "closed" (#281)
 * fixed #306: H&S null pointer exception when trying to place tile next to weird CCCF tile
 * fixed #304: Adjacent monastery and shrine should not challenge each if one of them is unoccupied
 * fixed #283: Sheep/wolf token counting error (panel should display remaining tokes not sheep sum)
 * fixed #295: fix NullPointerException at ClientMessageListener in error log when playing online
-
-
 
 ## 4.4.2
 *2020-04-07*